requirements.txt
django-rest-framework
djangorestframework-simplejwt
mysqlclient